# Boot performance tweaks