--- a/meta-balena-common/recipes-connectivity/openssh/balena-files/sshdgenkeys.service
+++ b/meta-balena-common/recipes-connectivity/openssh/balena-files/sshdgenkeys.service
@@ -1,8 +1,8 @@
 [Unit]
 Description=OpenSSH Key Generation
 RequiresMountsFor=/var /run
-Requires=resin-state.service bind-etc-ssh-hostkeys.service
-After=resin-state.service bind-etc-ssh-hostkeys.service
+Requires=mnt-state.mount bind-etc-ssh-hostkeys.service
+After=mnt-state.mount bind-etc-ssh-hostkeys.service
 
 [Service]
 ExecStart=@LIBEXECDIR@/sshd_check_keys