--- a/meta-balena-common/recipes-containers/balena-supervisor/balena-supervisor/migrate-supervisor-state.service
+++ b/meta-balena-common/recipes-containers/balena-supervisor/balena-supervisor/migrate-supervisor-state.service
@@ -1,7 +1,7 @@
 [Unit]
 Description=Balena supervisor state migrator
-Requires=resin-state.service
-After=resin-state.service
+Requires=mnt-state.mount
+After=mnt-state.mount
 ConditionPathExists=/mnt/state/root-overlay/etc/resin-supervisor/supervisor.conf
 ConditionPathExists=!/mnt/state/root-overlay/etc/balena-supervisor/supervisor.conf