--- a/meta-balena-common/recipes-containers/balena-supervisor/balena-supervisor/resin-data.mount
+++ b/meta-balena-common/recipes-containers/balena-supervisor/balena-supervisor/resin-data.mount
@@ -1,7 +1,7 @@
 [Unit]
 Description=resin-data bind mount
-Requires=resin-data.service balena-data-reset.service
-After=resin-data.service balena-data-reset.service
+Requires=mnt-data.mount balena-data-reset.service
+After=mnt-data.mount balena-data-reset.service
 
 [Mount]
 What=/mnt/data/resin-data