--- a/meta-balena-common/recipes-containers/balena/balena/var-lib-docker.mount
+++ b/meta-balena-common/recipes-containers/balena/balena/var-lib-docker.mount
@@ -1,7 +1,7 @@
 [Unit]
 Description=Docker data mountpoint
-Requires=resin-data.service balena-data-reset.service var-volatile-lib.service
-After=resin-data.service balena-data-reset.service var-volatile-lib.service
+Requires=mnt-data.mount balena-data-reset.service var-volatile-lib.service
+After=mnt-data.mount balena-data-reset.service var-volatile-lib.service
 
 [Mount]
 What=/mnt/data/docker