--- a/meta-balena-common/recipes-containers/hostapp-extensions-update/files/update-hostapp-extensions.service
+++ b/meta-balena-common/recipes-containers/hostapp-extensions-update/files/update-hostapp-extensions.service
@@ -1,7 +1,7 @@
 [Unit]
 Description=Hostapp extensions updater
-Requires=resin-data.service balena-data-reset.service balena.service
-After=resin-data.service balena-data-reset.service balena.service
+Requires=mnt-data.mount balena-data-reset.service balena.service
+After=mnt-data.mount balena-data-reset.service balena.service
 
 [Service]
 Type=oneshot