--- a/meta-balena-common/recipes-core/balena-rollback/files/rollback-altboot.service
+++ b/meta-balena-common/recipes-core/balena-rollback/files/rollback-altboot.service
@@ -17,7 +17,7 @@ Description=Balena rollback checks altboot
 DefaultDependencies=no
 Requires=resin-boot.service
 Before=balena-host.service balena.service
-After=resin-boot.service systemd-udev-settle.service
+After=resin-boot.service
 ConditionPathExists=/mnt/state/rollback-altboot-breadcrumb
 
 [Service]