--- a/meta-balena-common/recipes-core/balena-rollback/files/rollback-health.service
+++ b/meta-balena-common/recipes-core/balena-rollback/files/rollback-health.service
@@ -15,9 +15,9 @@
 [Unit]
 Description=Balena rollback checks health
 DefaultDependencies=no
-Requires=resin-boot.service mnt-sysroot-active.service
+Requires=resin-boot.service mnt-sysroot-active.mount
 Wants=openvpn.service balena.service
-After=resin-boot.service mnt-sysroot-active.service openvpn.service balena.service
+After=resin-boot.service mnt-sysroot-active.mount openvpn.service balena.service
 ConditionPathExists=/mnt/state/rollback-health-breadcrumb
 
 [Service]