--- a/meta-balena-common/recipes-core/systemd/timeinit/timeinit-rtc.service
+++ b/meta-balena-common/recipes-core/systemd/timeinit/timeinit-rtc.service
@@ -17,7 +17,6 @@ Description=Set the system clock from RTC
 ConditionVirtualization=!docker
 ConditionPathExists=/dev/rtc
 DefaultDependencies=no
-After=systemd-udev-settle.service
 Before=time-set.target
 
 [Service]