--- a/meta-balena-common/recipes-support/balena-config-vars/balena-config-vars/os-udevrules.service
+++ b/meta-balena-common/recipes-support/balena-config-vars/balena-config-vars/os-udevrules.service
@@ -1,7 +1,7 @@
 [Unit]
 Description=udev rules configuration from config.json
-Requires=resin-boot.service systemd-udev-settle.service bind-etc-udev-rules.d.service
-After=resin-boot.service systemd-udev-settle.service bind-etc-udev-rules.d.service
+Requires=resin-boot.service bind-etc-udev-rules.d.service
+After=resin-boot.service bind-etc-udev-rules.d.service
 
 [Service]
 Type=oneshot