--- a/meta-balena-common/recipes-support/balena-data-reset/balena-data-reset/balena-data-reset.service
+++ b/meta-balena-common/recipes-support/balena-data-reset/balena-data-reset/balena-data-reset.service
@@ -1,8 +1,8 @@
 [Unit]
 Description=balena reset OS data
 DefaultDependencies=no
-Requires=resin-data.service var-volatile.mount
-After=resin-data.service var-volatile.mount
+Requires=mnt-data.mount var-volatile.mount
+After=mnt-data.mount var-volatile.mount
 Before=umount.target
 Conflicts=umount.target
 ConditionPathExists=!@BALENA_DATA_MP@/remove_me_to_reset