--- a/meta-balena-common/recipes-support/balena-persistent-logs/balena-persistent-logs/balena-persistent-logs.service
+++ b/meta-balena-common/recipes-support/balena-persistent-logs/balena-persistent-logs/balena-persistent-logs.service
@@ -1,7 +1,7 @@
 [Unit]
 Description=Balena persistent logs
-Requires=resin-boot.service resin-state.service systemd-journald.service
-After=resin-boot.service resin-state.service systemd-journald.service
+Requires=resin-boot.service mnt-state.mount systemd-journald.service
+After=resin-boot.service mnt-state.mount systemd-journald.service
 
 [Service]
 Type=oneshot