--- a/meta-balena-common/recipes-support/resin-expand/resin-filesystem-expand/resin-filesystem-expand.service
+++ b/meta-balena-common/recipes-support/resin-expand/resin-filesystem-expand/resin-filesystem-expand.service
@@ -1,7 +1,7 @@
 [Unit]
 Description=Expand resin-data partition
 DefaultDependencies=no
-Before=resin-data.service umount.target
+Before=mnt-data.mount umount.target
 After=resin-boot.service
 Conflicts=umount.target
 ConditionVirtualization=!docker