--- a/meta-balena-common/recipes-support/resin-mounts/resin-mounts/etc-fake-hwclock.mount
+++ b/meta-balena-common/recipes-support/resin-mounts/resin-mounts/etc-fake-hwclock.mount
@@ -1,7 +1,7 @@
 [Unit]
 Description=Bind mount for /etc/fake-hwclock
-Requires=resin-state.service resin-state-reset.service
-After=resin-state.service resin-state-reset.service
+Requires=mnt-state.mount resin-state-reset.service
+After=mnt-state.mount resin-state-reset.service
 
 [Mount]
 What=/mnt/state/root-overlay/etc/fake-hwclock