--- a/meta-balena-common/recipes-support/resin-mounts/resin-mounts/mnt-data.automount
+++ b/meta-balena-common/recipes-support/resin-mounts/resin-mounts/mnt-data.automount
@@ -1,0 +1,9 @@
+[Unit]
+Description=mnt-data automount
+ConditionVirtualization=!docker
+
+[Automount]
+Where=/mnt/data
+
+[Install]
+WantedBy=multi-user.target