--- a/meta-balena-common/recipes-support/resin-mounts/resin-mounts/mnt-state.automount
+++ b/meta-balena-common/recipes-support/resin-mounts/resin-mounts/mnt-state.automount
@@ -1,0 +1,9 @@
+[Unit]
+Description=mnt-state automount
+ConditionVirtualization=!docker
+
+[Automount]
+Where=/mnt/state
+
+[Install]
+WantedBy=multi-user.target