--- a/meta-balena-common/recipes-support/resin-state-reset/resin-state-reset/resin-state-reset.service
+++ b/meta-balena-common/recipes-support/resin-state-reset/resin-state-reset/resin-state-reset.service
@@ -1,8 +1,8 @@
 [Unit]
 Description=Resin reset OS state
 DefaultDependencies=no
-Requires=resin-state.service var-volatile.mount
-After=resin-state.service var-volatile.mount
+Requires=mnt-state.mount var-volatile.mount
+After=mnt-state.mount var-volatile.mount
 Before=umount.target
 Conflicts=umount.target
 ConditionPathExists=!@BALENA_STATE_MP@/remove_me_to_reset