--- a/meta-balena-common/recipes-support/rng-tools/rng-tools_%.bbappend
+++ b/meta-balena-common/recipes-support/rng-tools/rng-tools_%.bbappend
@@ -1,0 +1,4 @@
+do_install:prepend() {
+        sed -i '/systemd-udev-settle/d' \
+            ${WORKDIR}/rngd.service
+}