--- a/meta-balena-kirkstone/recipes-core/plymouth/plymouth/0004-Avoid-depending-on-systemd-ask-password-path-unit.patch
+++ b/meta-balena-kirkstone/recipes-core/plymouth/plymouth/0004-Avoid-depending-on-systemd-ask-password-path-unit.patch
@@ -23,10 +23,10 @@ index 474667c..72ee09b 100644
  [Unit]
  Description=Show Plymouth Boot Screen
  DefaultDependencies=no
--Wants=systemd-ask-password-plymouth.path systemd-vconsole-setup.service systemd-udev-settle.service
--After=systemd-vconsole-setup.service systemd-udev-trigger.service systemd-udevd.service systemd-udev-settle.service resin-boot.service
-+Wants=systemd-vconsole-setup.service systemd-udev-settle.service
-+After=systemd-udev-trigger.service systemd-udevd.service systemd-udev-settle.service resin-boot.service
+-Wants=systemd-ask-password-plymouth.path systemd-vconsole-setup.service
+-After=systemd-vconsole-setup.service systemd-udev-trigger.service systemd-udevd.service resin-boot.service
++Wants=systemd-vconsole-setup.service
++After=systemd-udev-trigger.service systemd-udevd.service resin-boot.service
  Before=systemd-ask-password-plymouth.service
  Requires=resin-boot.service
  ConditionKernelCommandLine=!plymouth.enable=0