… meta-balena-common/recipes-support/resin-mounts/resin-mounts/resin-data.service
… meta-balena-common/recipes-support/resin-mounts/resin-mounts/resin-state.service