@alexgg 2023-02-21
@jakogut granted I am more conservative than you are, but there are a couple of things that makes me uneasy on this PR.

1) Removing the udev settle - common sense tell me that waiting on udev to finish on units that require udev rules to finish running make sense. So as I see it there a instability risk by introducing this change that is not well covered by the current automation testing.

2) Using systemd mounts - I am more comfortable with this change. However, I don't like that it's a partial change leaving the boot partition out. I'd rather just hold the PR until it also adds the boot partition.

Now, instead of making both changes at the same time I'd like to see them separate, and to see what is the individual boot time gain. Numbers for the RPI0 for example would maybe help too, as they could be bigger than in QEMU.

In summary, let's work on the remounts first, complete it, take the numbers and then decide whether to apply it. Once that has been merged and released and it has some time on the field to catch up unexpected issues, we can do the other improvement in the same way.