@jakogut 2023-02-21
Thanks for the conservative feedback, @alexgg. I felt the same way about both of your points, and wanted exposure and input on the PR as it stands. I'll revisit this later.