@jakogut 2023-02-21
@alexgg About waiting for `systemd-udev-settle`, the [docs](https://www.freedesktop.org/software/systemd/man/systemd-udev-settle.service.html) discourage its use, as hardware is enumerated and initialized asynchronously, and waiting for the event queue to be clear doesn't actually ensure all hardware has been initialized. I wrote more about this in the commit itself.

In the case of mounts units, the [docs](https://www.freedesktop.org/software/systemd/man/systemd.mount.html#Automatic%20Dependencies) say that block device backed filesystems automatically gain an implicit dependency on a device unit encapsulating the block device. This device unit ensures the device is initialized and the device node is present before the mount unit is started. This allows for mounts to be done concurrently and asynchronously, rather than stalling while all udev events are processed.

I'm open to ideas for improving our test suite to ensure changes like this aren't breaking, as I'm also not 100% sure they're safe.