@alexgg 2023-02-21
> @alexgg About waiting for `systemd-udev-settle`, the [docs](https://www.freedesktop.org/software/systemd/man/systemd-udev-settle.service.html) discourage its use, as hardware is enumerated and initialized asynchronously, and waiting for the event queue to be clear doesn't actually ensure all hardware has been initialized. I wrote more about this in the commit itself.

As I understand the docs, waiting to settle means that you have processed all the event queue that has been received, while you may receive new events that have not been yet queued after the settle.

Not waiting for the settle means the system would miss on events in the queue. We wait because if the disks are not processed the boot halts, and we have seen race conditions in the past. So it's a trade off between boot performance and boot reliability.  That's why knowing exactly how much quicker it would boot on slower devices is a factor to consider.

> 
> In the case of mount units, the [docs](https://www.freedesktop.org/software/systemd/man/systemd.mount.html#Automatic%20Dependencies) say that block device backed filesystems automatically gain an implicit dependency on a device unit encapsulating the block device. This device unit ensures the device is initialized and the device node is present before the mount unit is started. This allows for mounts to be done concurrently and asynchronously, rather than stalling while all udev events are processed.
> 

About the mount units, I do favor it too, it just needs to work also for the boot partition not to use different mechanisms for consistency.